Backend-Django/requirement.txt
appdirs==1.4.4
asgiref==3.3.1
cffi==1.14.4
cryptography==3.3.1
distlib==0.3.1
Django==3.1.5
django-autoconfig==0.8.0
django-closuretree==1.2.0
django-dbbackup==3.3.0
django-fernet-fields==0.6
django-js-asset==1.2.2
django-mptt==0.12.0
django-treenode==0.15.0
django-unixtimestampfield==0.4.0
Django-Verify-Email==0.0.5
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
djangorestframework-recursive==0.1.2
djongo==1.3.3
filelock==3.0.12
inflection==0.5.1
mod-wsgi==4.7.1
pdfminer3k==1.3.4
Pillow==8.1.0
ply==3.11
pycparser==2.20
pymongo==3.11.2
PyPDF2==1.26.0
pytz==2020.5
six==1.15.0
slate3k==0.5.3
sqlparse==0.2.4
virtualenv==20.4.2
